Split parsing outside of expression.rs
